Move Types and Protocol to Api
