+++ marlowe-chain-sync/marlowe-chain-sync.cabal
@@ -60,10 +60,9 @@ library
     Language.Marlowe.Runtime.ChainSync.Database.PostgreSQL
     Language.Marlowe.Runtime.ChainSync.Genesis
     Language.Marlowe.Runtime.ChainSync.NodeClient
-    Language.Marlowe.Runtime.ChainSync.Protocol
     Language.Marlowe.Runtime.ChainSync.Server
     Language.Marlowe.Runtime.ChainSync.Store
-    Language.Marlowe.Runtime.ChainSync.Types
+    Language.Marlowe.Runtime.ChainSync.Api
   other-modules:
   build-depends:
       base >= 4.9 && < 5
